+++ packages/Core/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed issue with `create table` query.
 - `config` option is now optional.
 - Added method to list databases.
+- Fixed issue with nested function calls.
 
 ## [2.2.7] 4-19-2019
 - Fixed an issue with select table code.